Improve VSCode Extension - Alpha Version: Server Starter, Projection View, ... (#102)
* Add scripts to copy servers to local folder

* Start GLSP-Server from vscode extension

* Start Model-Server from vscode extension

* Load theme css in glsp not theia integration

* Use action based approach for theme

* Use theme in vscode

* Use esbuild script file

* Fix redo and undo handling

* Use glsp projection view for graph

* Use bold and italic for svg header

* Change uml to biguml as prefix for vscode extension

* Use icons for the .uml and .unotation file extensions

* Use random port in production

* Fix font error log in console

* Use label edit validator for validating blank inputs and multiplicity

* Update icon colors

* Add workaround for java 17 gson support (e.g., flag --add-opens)

* Fix style loading and make the height of the tool palette dynamic

* Log vscode production server logs to output channel

* Make property not moveable and remove iconlabelcompartment

* Fix build process and server killing for windows

* Small fixes

* Use multistep wizard for creating files

* Use polyline instead of manhatten for routing

* Allow to create uml diagrams from the context menu

* Make labels align to the edge

* Enable grid snapper

* Create release documentation for sharing and fix reference deletion for operation

* Initialize settings to hide unotation files

* Fix settings and new file wizard validation

* Improve new file input name validation

* Small fixes for readme and concerning when the biguml commands are shown

* Use zip as archive

+++ server/com.eclipsesource.uml.modelserver/src/main/java/com/eclipsesource/uml/modelserver/uml/diagram/class_diagram/reference/OperationReferenceRemover.java
@@ -26,6 +26,10 @@ public final class OperationReferenceRemover extends BaseCrossReferenceProcessor
    @Override
    protected List<Command> process(final ModelContext context, final Setting setting, final Operation self,
       final EObject interest) {
-      return List.of(new DeleteOperationSemanticCommand(context, self));
+      if (self.getOwner().equals(interest)) {
+         return List.of(new DeleteOperationSemanticCommand(context, self));
+      }
+
+      return List.of();
    }
 }
